requirements.txt
Django==1.8.5
pep8==1.6.2
djangorestframework==3.0.0
drf-nested-routers==0.9.0